refactor(random): simplify shuffle

+++ src/misc/random.h
@@ -14,6 +14,6 @@ void seedRNG(RNG* rng, uint64_t init_state, uint64_t init_seq);
 
 uint8_t generateBoundedRandomNumber(RNG* rng, uint8_t bound);
 
-void shuffle(int* array, size_t n, RNG* rng);
+void shuffle(int* array, int n, RNG* rng);
 
 #endif //UTTT2_RANDOM_H
